Separate creation tests for user_bonuses and task_suites
